Add interface for non-file LOAD DATA LOCAL INFILE (#xx)
LOAD DATA LOCAL INFILE is a rapid load mechanism in MySQL. ETL,
machine learning, and specialized accelerators are some of the
reasons that a rapid generation of data can occur. These are
application domains where Python is frequently used. Placing
generated data into to a file before a load is an additional
complication. This commit prevents a temporary file generation.

In this commit the load_infile argument is expanded from a
boolean into a dictionary. In true pythonic style, a dictionary
implements `__getitem__` and during the client side LOAD DATA
LOCAL INFILE implementation it is the filename that is looked
up in the dictionary.

Like file objects, a python `with` syntax (PEP-343)
is used on the object retrieved from the dictionary. A conformant
`__enter__`/`__exit__` methods is required to be implemented. The
object returned from __enter__, will have have its `read` method
called to retrieve data as if it was the file being uploaded.
The `read` method will return CSV, or otherwise separated, data
as bytes. When the `read` method is complete it will return None.